refactor(embed): share submission-error handling across hosted-web flows
Extends the biometric-kyc fix to every hosted-web product flow and removes
the duplicated per-entry-point copies by moving the logic into a shared
submission-error.js module used by biometric-kyc, ekyc, basic-kyc,
doc-verification, enhanced-document-verification and smartselfie-auth.

- displayErrorMessage now reuses a single #submission-error-message element
  so repeated failed submissions update the message in place instead of
  stacking a new one per click. Guards against a missing <main> so it can
  never mask the original error inside a catch handler.
- Network-level failures (offline / timeout) show the actionable
  pages.error.checkInternet copy instead of the generic message, via a
  shared submissionErrorMessage() helper. The isNetworkError cause-chain
  walk is depth-capped so a circular cause chain can't loop forever.

Addresses the automated reviewer feedback (cycle guard + null-main guard).
e-signature is left unchanged — its only displayErrorMessage use is a
checkbox validation already cleared by resetForm(), so it has no stacking bug.
