+++ packages/embed/src/js/doc-verification.js
@@ -19,6 +19,10 @@ import {
   idInfoToIdSelection,
 } from './id-info-utils.js';
 import { fetchWithTimeout } from './fetch-with-retry.js';
+import {
+  displayErrorMessage,
+  submissionErrorMessage,
+} from './submission-error.js';
 import initIframeSentry from './sentry-iframe-init.js';
 import { createDocSubmission } from './doc-submission.js';
 
@@ -804,23 +808,11 @@ window.Sentry = Sentry;
       event.target.disabled = false;
     } catch (error) {
       event.target.disabled = false;
-      displayErrorMessage(t('pages.error.generic'));
+      displayErrorMessage(submissionErrorMessage(error, t));
       console.error(
         `SmileIdentity - ${error.name || error.message}: ${error.cause}`,
       );
     }
-  }
-
-  function displayErrorMessage(message) {
-    const p = document.createElement('p');
-
-    p.textContent = message;
-    p.classList.add('validation-message');
-    p.style.fontSize = '1.5rem';
-    p.style.textAlign = 'center';
-
-    const main = document.querySelector('main');
-    main.prepend(p);
   }
 
   async function createZip() {
