+++ packages/embed/src/js/ekyc.js
@@ -23,6 +23,10 @@ import {
   captureInitApiFailure,
 } from './init-api-sentry.js';
 import { fetchWithTimeout } from './fetch-with-retry.js';
+import {
+  displayErrorMessage,
+  submissionErrorMessage,
+} from './submission-error.js';
 import initIframeSentry from './sentry-iframe-init.js';
 
 initIframeSentry('ekyc');
@@ -1015,23 +1019,11 @@ initIframeSentry('ekyc');
       complete();
     } catch (error) {
       if (event && event.target) event.target.disabled = false;
-      displayErrorMessage('Something went wrong');
+      displayErrorMessage(submissionErrorMessage(error, translate));
       console.error(
         `SmileIdentity - ${error.name || error.message}: ${error.cause}`,
       );
     }
-  }
-
-  function displayErrorMessage(message) {
-    const p = document.createElement('p');
-
-    p.textContent = message;
-    p.classList.add('validation-message');
-    p.style.fontSize = '1.5rem';
-    p.style.textAlign = 'center';
-
-    const main = document.querySelector('main');
-    main.prepend(p);
   }
 
   async function submitIdInfoForm() {
